requirements.txt
pandas
scikit-learn
networkx
psycopg2-binary
python-dotenv
